# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CDNFoundry Control Plane API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/admin/audit-logs": {
            "get": {
                "operationId": "audit.log.controller",
                "tags": [
                    "Administrator"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
        "/admin/domains/{domain}/force-verify": {
            "post": {
                "operationId": "domain.verification.controller",
                "tags": [
                    "Administrator"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
        "/nameservers": {
            "get": {
                "operationId": "nameserver.controller",
                "tags": [
                    "Account"
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "responses": {
            "ValidationError": {
                "description": "The provided data is invalid.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            }
        },
        "schemas": {
            "ValidationError": {
                "type": "object",
                "required": [
                    "message",
                    "code",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string",
                        "const": "validation_failed"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}